refactoring/documenting TLS classes, refs#71
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@7074 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/traffic_lights/MSActuatedTrafficLightLogic.cpp
@@ -51,7 +51,7 @@ MSActuatedTrafficLightLogic::MSActuatedTrafficLightLogic(
     MSNet &net, MSTLLogicControl &tlcontrol,
     const std::string &id, const std::string &subid,
     const Phases &phases,
-    size_t step, SUMOTime delay, SUMOReal maxGap, SUMOReal passingTime,
+    unsigned int step, SUMOTime delay, SUMOReal maxGap, SUMOReal passingTime,
     SUMOReal detectorGap)
         : MSSimpleTrafficLightLogic(net, tlcontrol, id, subid, phases, step, delay),
         myContinue(false),
@@ -145,7 +145,7 @@ MSActuatedTrafficLightLogic::duration() const {
     // define the duration depending from the number of waiting vehicles of the actual phase
     int newduration = currentPhaseDef()->minDuration;
     const std::bitset<64> &isgreen = currentPhaseDef()->getDriveMask();
-    for (size_t i=0; i<isgreen.size(); i++)  {
+    for (unsigned int i=0; i<(unsigned int) isgreen.size(); i++)  {
         if (isgreen.test(i))  {
             const std::vector<MSLane*> &lanes = getLanesAt(i);
             if (lanes.empty())    {
@@ -235,7 +235,7 @@ MSActuatedTrafficLightLogic::gapControl() {
 
     // now the gapcontrol starts
     const std::bitset<64> &isgreen = currentPhaseDef()->getDriveMask();
-    for (size_t i=0; i<isgreen.size(); i++)  {
+    for (unsigned int i=0; i<(unsigned int) isgreen.size(); i++)  {
         if (isgreen.test(i))  {
             const std::vector<MSLane*> &lanes = getLanesAt(i);
             if (lanes.empty())    {
